+++ app/services/passmeister_service.rb
@@ -4,19 +4,25 @@ require 'json'
 class PassmeisterService
   class Error < StandardError; end
 
-  TEMPLATE_ID = 'P428195'.freeze
+  PASS_TYPE_ID = 'P428195'.freeze
   API_BASE = 'https://www.passmeister.com/api/v1'.freeze
+  LIZENZCHECK_URL = 'https://sr.floorball.de/lizenzcheck/'.freeze
 
   def self.create_or_update_pass(referee)
+    pass_id = "referee-#{referee.lizenznummer}"
+    club_name = referee.club&.name.to_s
+    barcode_link = "#{LIZENZCHECK_URL}?#{URI.encode_www_form(q: referee.lizenznummer)}"
+
     uri = URI("#{API_BASE}/pass")
+    uri.query = URI.encode_www_form(passTypeId: PASS_TYPE_ID, passId: pass_id)
+
     payload = {
-      'passId'         => "referee-#{referee.lizenznummer}",
-      'templateId'     => TEMPLATE_ID,
-      'memberName'     => { 'value' => "#{referee.vorname} #{referee.nachname}" },
-      'club'           => { 'value' => referee.club&.name.to_s },
-      'memberNumber'   => { 'value' => referee.lizenznummer.to_s },
-      'barcode'        => { 'label' => referee.lizenznummer.to_s },
-      'expirationDate' => next_rotation_expiry.iso8601
+      'field.memberName.value'   => "#{referee.vorname} #{referee.nachname}",
+      'field.memberNumber.value' => referee.lizenznummer.to_s,
+      'field.club.value.de'      => club_name,
+      'field.club.value.en'      => club_name,
+      'field.barcode.label'      => barcode_link,
+      'expiresAt'                => next_rotation_expiry.utc.iso8601
     }
 
     req = Net::HTTP::Post.new(uri)
@@ -28,7 +34,9 @@ class PassmeisterService
       http.request(req)
     end
 
-    raise Error, "Passmeister-Fehler #{response.code}: #{response.body}" unless response.is_a?(Net::HTTPSuccess)
+    unless response.is_a?(Net::HTTPSuccess)
+      raise Error, "Passmeister-Fehler #{response.code}: #{response.body.to_s[0, 500]}"
+    end
 
     JSON.parse(response.body)
   rescue SocketError, Errno::ECONNREFUSED, Errno::ETIMEDOUT, Errno::ECONNRESET,
@@ -42,7 +50,7 @@ class PassmeisterService
     base = 2030
     target = base
     target += 4 while target <= Date.today.year
-    Date.new(target, 6, 30)
+    Time.utc(target, 6, 30, 23, 59, 59)
   end
 
   def self.api_key